Config map update consistency
If an update of the configuration map fails while the plugin is running, it
keeps the old config map value.  This is inconsistent with the behavior of a
rolling update since the restarting plugin would not provide any configuration
sets.

Fix the code to always provide either the newest configuration set or none at
all.  This way, dynamic updates and rolling updates should behave the same
way.

Signed-off-by: Juergen Christ <jchrist@linux.ibm.com>