Fix empty machineid comparison
The main function fetched the machineid into a new constructed
local variable instead of the global variable. So the global
variable was still empty and thus the machineid matching did
not work. So now the new global variable MachineId really
contains the machine id of the machine where the cex plugin
instance is running and APQNs with machineid matching works.

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>