Reannounce the list of crypto config sets in case of changes
As of now changes in the crypto config sets are not
handled. So if the configuration gets an additional
config set or a config set is removed the plugin has
no awareness about this. As a result, the plugin daemonset
needs to be restarted to handle this type of configuration
change.

This patch introduces the correct handling of config
set changes within a configuration. So now an additional
plugin instance is started when a new config set comes
into existence or a plugin instance is shut down when a
config set vanishes.

Not new but worth mentioning: If an invalid configuration is
given, all plugin instances are shut down and a message
  Plugin: No crypto config sets available, check configuration !
appears every 120s in the log together with information about
why the configuration is not valid. If later a valid
configuration is available the plugin instances are driven up
again and everything works as expected.

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>