Make CryptoConfig methods handle nil object pointer correctly
Make the functions
GetListOfSetNames(), GetCryptoConfigSet() and GetCryptoConfigSetForThisAPQN()
save for getting a nil object pointer which may be caused by an invalid
new configuration.

This patch also fixes some minor details. When a error
log string is generated for further use by the caller it must
not end with a newline because the caller does a printf
with "...error: %s...\n", err) which causes double newlines
in the log.

And it fixes a type: manufactor... -> manufactur...

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>