Introduce metrics collector for the cex device plugin
This patch introduces a new component within the cex plugin code:
A metrics collector which provides an interface for the other
components to report some important notifications to the collector.
This data is then pushed to the prometheus exporter pot running
somewhere in the cluster.

The values to reach the cex prometheus exporter are given
via environment variables:

  CEX_PROM_EXPORTER_COLLECTOR_SERVICE gives the service name
    and defaults to "cex-prometheus-exporter-collector-service"
  CEX_PROM_EXPORTER_COLLECTOR_SERVICE_NAMESPACE is the namespace
    where the exporter lives. Default is empty as it is assumed that
    the cex device plugin runs in the same namespaces as the cex
    prometheus exporter.
  CEX_PROM_EXPORTER_COLLECTOR_SERVICE_PORT gives the port number
    for the exporter collector port and defaults to 12358

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>