Introduce prometheus exporter for the cex device plugin
This is the first version of the cex prometheus exporter.
That is the mediator between the prometheus server and
the cex plugin daemonset instances running on the worker
nodes.

The prometheus client port where prometheus pulls
the metrics from is given as environment variable
PROMETHEUS_SERVICE_PORT (default value 9939).
This port needs to be exported and a prometheus
service monitor needs to be declared also, so that
the prometheus server can pull the metrics data.

The collector listener port where all the cex plugin
instances push their metrics data in is given as
environment variable COLLECTOR_SERVICE_PORT (default
value 12358). This port needs to be exported as a clusterwide
service with name 'cex-prometheus-exporter-collector-service'
in namespace 'cex-device-plugin' so that the cex plugin instances
are able to tcp connect to the cex prometheus exporter.

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>