Fix crash on empty config set
The plugin crashed over a nil pointer when the crypto configuration
was updated and an existing configset removed:
- Reworked the way how Livesysfs is parsed
- Reworked the way how Overcommit is parsed
As both parameters are optional for an crypto config set, now both are
parsed with the help of an intermediate field (_livesysfs and
_overcommit) and after parsing the real fields Livesysfs and
Overcommit get either the parsed value or a -1 (indicating there was
no parameter given). So there is no pointer checking needed any more
in the other code using these fields.
- Add a check for nil pointer in method makePluginDevsFromAPQNs(). So
  this function now simple returns an empty list of devices if the
  config set is nil.
- During testing it has been shown that the comparison between current
  and new config set needs rework. It always found a change because a
  non given overcommit or livesysfs value was compared against the
  default value which gives a hit as -1 is always not equal to 0 or 1.
  So now first the defaults are applied to the new config set and then
  it is compared against the existing config set.

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>