Fix crash caused by rase condition during shutdown of an plugin instance
There have been occasional crashes caused by use of an already closed
channel. On shutdown of a plugin instance the Stop() method closes the
chanced channel whereas it is possible that a go routine concurrently
is about to check if something has changed and if something has
changed pushes an empty item into the changed channel - which may
already have been closed.

To fix this a WaitGroup variable is now part of the Plugin struct
counting the number of active users for this changed channel. So now
the procedure to gracefully shut down a Plugin instance is to first
signal the shutdown via stop channel and then wait on the WaitGroup
until the counter of the chanced channel drops to 0. Then the chanced
channel can be safely closed.

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>